Add Color.Combination and begin to use it in a few places
There's only six colors. No reason we can't enumerate every possible combination thereof. This makes some of the manipulation more pleasant and guarantees the operations are efficient. It also provides a useful primitive for clients of the library.

I'm trying to avoid having any API be required to return a Color.Combination, but I can't promise it'll always be that way.

Signed-off-by: Emi <dealuminis@gmail.com>