Add utility to determine land color identity from type line
It's not a direct member of TypeLine, but a static one. The Scryfall impl is going to use this to correctly assign the color identity of lands. I don't love using strings here, but that's subtypes for you -- maybe in the future there'll be a LandType enum that can be checked against type lines, but for now this will do.

Signed-off-by: Emi <dealuminis@gmail.com>