Embed some geometric information for printed faces
Basically, printed faces now have `onBack()` which returns true if that face is printed on the back side of the piece of cardboard (where the standard magic card back usually is), and `frame()` which returns information about the location and orientation of that face on that side of the cardboard (the bounding rectangle and rotation-required-to-upright). I also defined a big set of standard card frames that encompasses basically all of usual MtG, except for Who/What/When/Where/Why (hate that card, although this could technically do it...).

The only real thing I hate about this is meld. Dang meld cards. >:(

Signed-off-by: Emi <dealuminis@gmail.com>