Add Card.Printing.Reference and DataSource.printing(Reference)
I was doing this a lot in the deckbuilder, then started doing it in the Scryfall plugin, and finally realized that this is a pretty standardized thing and should be available more generally. It also makes sense to add a lookup method to the data source based on printing references.

Signed-off-by: Emi <dealuminis@gmail.com>